Add PSI (pitching skill index); split sabermetrics from adjustment
Mirror of BSI for pitching: PSI = run-prevention index adjusted for the
offensive quality of the batters faced. Skill is measured by runs allowed;
opponent offense by wOBA, which a reliability experiment showed is a
cleaner skill signal than runs scored (more reliable and more predictive
of future run production).

- Replace ERA- with PSI on pstat_user, pstat_user_char, and the pstat_char
  leaderboard (ranked by PSI); pstat_all shows the unadjusted index.
- Pitching baselines now cache full PitchingStats (PSI needs runs/BF).
- New by-user batting table for opponent offensive quality.

Reorganize the pure math out of the overgrown helpers/wrc.py:
- helpers/sabermetrics.py: metric definitions (wOBA, wRC+, pitching index,
  league runs/PA).
- helpers/opponent_adjustment.py: the strength-of-schedule framework
  (schedules, shrinkage, BSI + PSI adjustments).
- Tests split to match (test_sabermetrics + test_opponent_adjustment); 69 total.
