+++ main/resources/stat_constants.py
@@ -1,5 +1,7 @@
-"""Constants for advanced batting metrics (wOBA / wRC+) and the opponent-quality
-(strength-of-schedule) adjustment. The calculation logic lives in helpers/wrc.py."""
+"""Constants for the advanced batting/pitching metrics (wOBA / wRC+ / pitching
+index) and the opponent-quality (strength-of-schedule) adjustment. The metric
+logic lives in helpers/sabermetrics.py; the adjustment in
+helpers/opponent_adjustment.py."""
 
 # Linear weights for wOBA. Derived from regression on Project Rio game data
 # (they match the values in RioStatParsing's fip.py / xba.py), NOT MLB weights.
@@ -14,13 +16,13 @@ LINEAR_WEIGHTS = {
 
 # League runs per plate appearance. Used as the wRC+ denominator and as a fallback
 # when a mode's pitching table is empty; normally derived live per mode from
-# total runs allowed / total batters faced (see helpers/wrc.league_runs_per_pa).
+# total runs allowed / total batters faced (see helpers/sabermetrics.league_runs_per_pa).
 DEFAULT_LEAGUE_RUNS_PER_PA = 0.16961480888026817
 
 # --- Opponent-quality adjustment tuning -------------------------------------
 # The adjustment is additive on the run scale: a batter who faced tougher-than-
 # average run prevention gets a higher adjusted wRC+. Two regression layers keep
-# small samples from dominating. See helpers/wrc.py for the model.
+# small samples from dominating. See helpers/opponent_adjustment.py for the model.
 
 # Transfer coefficient. 1.0 = full additive transfer (justified by the run-scale
 # model); lower values dampen the adjustment uniformly.
@@ -29,6 +31,10 @@ OPPONENT_TRANSFER_COEFF = 1.0
 # Per-opponent reliability: the batters-faced count at which an opponent's
 # run-prevention rate is trusted halfway. Below this it's regressed toward league.
 OPPONENT_REGRESSION_BF = 100
+
+# Same idea on the offensive side (for the pitching adjustment): the plate-
+# appearance count at which an opponent's offensive rate is trusted halfway.
+OPPONENT_REGRESSION_PA = 100
 
 # Schedule reliability: the number of games at which a batter's strength-of-
 # schedule estimate is trusted halfway. Thin schedules are damped toward zero.
